Clean up basic components - move basic components to basic folder - update basic components imports - remove unused imports
