+++ components/week/CloseWeekOverviewButton.tsx
@@ -1,4 +1,4 @@
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import React from 'react';
 import { colors } from '../../assets/globalStyling/colors';
 import PressableIcon from '../icons/PressableIcon';
